% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_countryname.R
\name{code_countryabbrv}
\alias{code_countryabbrv}
\title{Code Abbreviations of Country Names}
\usage{
code_countryabbrv(name)
}
\arguments{
\item{name}{}
}
\value{
A character vector of an abbreviation of specified country name
}
\description{
Insert abbreviations of country names.
}
\examples{
name <- c("argentine", "albanie", "Bélarus")
code_countryname(name)
}
