% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_countryname.R
\name{code_countryname}
\alias{code_countryname}
\title{Code Country Names}
\usage{
code_countryname(name, abbrev = FALSE)
}
\arguments{
\item{name}{A character vector of country name}

\item{abbrev}{Do you want 3 letter country
abbreviations to be returned?
False by default.}
}
\value{
A character vector of country name
}
\description{
Standardize spellings for country names,
as well as country names in other languages.
}
\examples{
name <- c("argentine", "albanie", "Bélarus")
code_countryname(name)
code_countryname(name, abbrev = TRUE)
}
