% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qTrade-references.R
\docType{data}
\name{references}
\alias{references}
\title{references database documentation}
\format{
The references database is a list that contains the
following 1 datasets: DESTA_REF.
For more information and references to each of the datasets used,
please use the \code{data_source()}, \code{data_contrast()}, and \code{data_evolution()}
functions.
\describe{
\item{DESTA_REF: }{A dataset with 1000 observations and the following
3 variables: qID1, RefType, qID2.}
}
}
\usage{
references
}
\description{
references database documentation
}
\keyword{datasets}
