% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumNA.R
\name{sumNA}
\alias{sumNA}
\title{Sum NAs}
\usage{
sumNA(data)
}
\arguments{
\item{data}{name of dataset}
}
\value{
A vector with the sum of 'NA's for each row
}
\description{
Sums the number of NAs across the columns for each row in the dataset
}
\examples{
data <- data.frame(a = c("a", NA),
b = c(NA, NA))
data$missing <- sumNA(data)
}
