% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-texts.R
\docType{data}
\name{texts}
\alias{texts}
\title{texts database documentation}
\format{
The texts database is a list that contains the
following 2 datasets: AGR_TXT, TOTA_TXT.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{AGR_TXT: }{A dataset with 352 observations and the following
3 variables: Title, Beg, TreatyText.}
\item{TOTA_TXT: }{A dataset with 450 observations and the following
2 variables: ID, TreatyText.}
}
}
\usage{
texts
}
\description{
texts database documentation
}
\details{
\subsection{NA}{

NA
}
}
\keyword{datasets}
