% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qTrade-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements database documentation}
\format{
The agreements database is a list that contains the
following 4 datasets: DESTA, TREND, LABPTA, GPTAD.
For more information and references to each of the datasets used,
please use the \code{data_source()}, \code{data_contrast()}, and \code{data_evolution()}
functions.
\describe{
\item{DESTA: }{A dataset with 1217 observations and the following
7 variables: Title, Beg, Signature, Force, Agreement type, Document type, qID.}
\item{TREND: }{A dataset with 730 observations and the following
6 variables: TREND_ID, Title, Beg, Signature, Force, qID.}
\item{LABPTA: }{A dataset with 483 observations and the following
6 variables: LABPTA_ID, Title, Beg, Signature, Force, qID.}
\item{GPTAD: }{A dataset with 352 observations and the following
8 variables: Title, Beg, Signature, Force, Agreement type, Document type, WTO, qID.}
}
}
\usage{
agreements
}
\description{
agreements database documentation
}
\keyword{datasets}
