% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qTrade-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{Agreements database documentation}
\format{
The agreements database is a list that contains the
following 4 datasets: DESTA, TREND, LABPTA, GPTAD.
For more information and references to each of the datasets used,
please use the \code{data_source()}, \code{data_contrast()}, and \code{data_evolution()}
functions.
\describe{
\item{DESTA: }{A dataset with 987 observations and the following
10 variables: DESTA_ID, Title, Beg, Signature, Force, D, L, J, WTO, qID.}
\item{TREND: }{A dataset with 730 observations and the following
6 variables: TREND_ID, Title, Beg, Signature, Force, qID.}
\item{LABPTA: }{A dataset with 483 observations and the following
6 variables: LABPTA_ID, Title, Beg, Signature, Force, qID.}
\item{GPTAD: }{A dataset with 346 observations and the following
9 variables: GPTAD_ID, Title, Beg, Signature, Force, D, L, WTO, qID.}
}
}
\usage{
agreements
}
\description{
Agreements database documentation
}
\keyword{datasets}
