% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements database documentation}
\format{
The agreements database is a list that contains the
following 7 datasets: DESTA, GPTAD, LABPTA, TOTA, TREND, HUGGO, AIGGO.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{DESTA: }{A dataset with 959 observations and the following
10 variables: manyID, treatyID, Title, Begin, AgreementType, DocType,
GeogArea, Signature, Force, destaID.}
\item{GPTAD: }{A dataset with 340 observations and the following
10 variables: manyID, treatyID, Title, Begin, AgreementType, DocType,
GeogArea, Signature, Force, gptadID.}
\item{LABPTA: }{A dataset with 483 observations and the following
7 variables: manyID, treatyID, Title, Begin, Signature, Force, labptaID.}
\item{TOTA: }{A dataset with 442 observations and the following
7 variables: manyID, treatyID, Title, Begin, Signature, Force, totaID.}
\item{TREND: }{A dataset with 710 observations and the following
7 variables: manyID, treatyID, Title, Begin, Signature, Force, trendID.}
\item{HUGGO: }{A dataset with 548 observations and the following
10 variables: manyID, treatyID, Title, Begin, Signature, Force, End, url,
TreatyTextStatus, Coder.}
\item{AIGGO: }{A dataset with 1389 observations and the following
8 variables: manyID, treatyID, Title, Begin, Signature, Force, accessionC,
accessionP.}
}
}
\usage{
agreements
}
\description{
agreements database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $DESTA
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             959|     0     |       0       |
#> |   treatyID    |  character  |             959|     0     |       0       |
#> |     Title     |  character  |             959|     0     |       0       |
#> |     Begin     |    mdate    |             959|    81     |     8.45      |
#> | AgreementType |  character  |             959|     0     |       0       |
#> |    DocType    |  character  |             959|     0     |       0       |
#> |   GeogArea    |  character  |             959|     0     |       0       |
#> |   Signature   |    mdate    |             959|    81     |     8.45      |
#> |     Force     |    mdate    |             959|    242    |     25.23     |
#> |    destaID    |  character  |             959|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $GPTAD
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             340|     0     |       0       |
#> |   treatyID    |  character  |             340|     0     |       0       |
#> |     Title     |  character  |             340|     0     |       0       |
#> |     Begin     |    mdate    |             340|    30     |     8.82      |
#> | AgreementType |  character  |             340|     0     |       0       |
#> |    DocType    |  character  |             340|     0     |       0       |
#> |   GeogArea    |  character  |             340|    75     |     22.06     |
#> |   Signature   |    mdate    |             340|    53     |     15.59     |
#> |     Force     |    mdate    |             340|    63     |     18.53     |
#> |    gptadID    |  character  |             340|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $LABPTA
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             483|     0     |       0       |
#> |   treatyID    |  character  |             483|     0     |       0       |
#> |     Title     |  character  |             483|     0     |       0       |
#> |     Begin     |    mdate    |             483|     0     |       0       |
#> |   Signature   |    mdate    |             483|     0     |       0       |
#> |     Force     |    mdate    |             483|     0     |       0       |
#> |   labptaID    |  character  |             483|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $TOTA
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             442|     0     |       0       |
#> |   treatyID    |  character  |             442|     0     |       0       |
#> |     Title     |  character  |             442|     0     |       0       |
#> |     Begin     |    mdate    |             442|     0     |       0       |
#> |   Signature   |    mdate    |             442|     0     |       0       |
#> |     Force     |    mdate    |             442|     3     |     0.68      |
#> |    totaID     |  character  |             442|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $TREND
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             710|     0     |       0       |
#> |   treatyID    |  character  |             710|     0     |       0       |
#> |     Title     |  character  |             710|     0     |       0       |
#> |     Begin     |    mdate    |             710|     5     |      0.7      |
#> |   Signature   |    mdate    |             710|     5     |      0.7      |
#> |     Force     |    mdate    |             710|     5     |      0.7      |
#> |    trendID    |  character  |             710|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO
#> -----------------------------------------------------------------------------
#> |  Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> -----------------------------------------------------------------------------
#> |     manyID     |  character  |             548|     0     |       0       |
#> |    treatyID    |  character  |             548|     0     |       0       |
#> |     Title      |  character  |             548|     0     |       0       |
#> |     Begin      |    mdate    |             548|     0     |       0       |
#> |   Signature    |    mdate    |             548|     0     |       0       |
#> |     Force      |    mdate    |             548|     8     |     1.46      |
#> |      End       |    mdate    |             548|    492    |     89.78     |
#> |      url       |  character  |             548|     0     |       0       |
#> |TreatyTextStatus|  character  |             548|     0     |       0       |
#> |     Coder      |  character  |             548|     0     |       0       |
#> -----------------------------------------------------------------------------
#> 
#> 
#> $AIGGO
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            1389|     0     |       0       |
#> |   treatyID    |  character  |            1389|     0     |       0       |
#> |     Title     |  character  |            1389|     0     |       0       |
#> |     Begin     |    mdate    |            1389|    115    |     8.28      |
#> |   Signature   |    mdate    |            1389|    137    |     9.86      |
#> |     Force     |    mdate    |            1389|    252    |     18.14     |
#> |  accessionC   |  character  |            1389|   1000    |     71.99     |
#> |  accessionP   |  character  |            1389|   1139    |      82       |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
