% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements database documentation}
\format{
The agreements database is a list that contains the
following 4 datasets: DESTA, TREND, LABPTA, GPTAD.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{DESTA: }{A dataset with 987 observations and the following
11 variables: many_ID, Title, Beg, D, L, J, Signature, Force, treaty_ID, DESTA_ID, WTO.}
\item{TREND: }{A dataset with 730 observations and the following
7 variables: many_ID, Title, Beg, Signature, Force, treaty_ID, TREND_ID.}
\item{LABPTA: }{A dataset with 483 observations and the following
7 variables: many_ID, Title, Beg, Signature, Force, treaty_ID, LABPTA_ID.}
\item{GPTAD: }{A dataset with 346 observations and the following
9 variables: many_ID, Title, Beg, D, L, Signature, Force, treaty_ID, GPTAD_ID.}
}
}
\usage{
agreements
}
\description{
agreements database documentation
}
\details{
\preformatted{## $DESTA
## -- Data Summary ------------------------
##                            Values
## Name                       X[[i]]
## Number of rows             987   
## Number of columns          11    
## _______________________          
## Column type frequency:           
##   character                11    
## ________________________         
## Group variables            None  
## 
## -- Variable type: character --------------------------------------------------------------------------------------------------------------------------------------------------------
## # A tibble: 11 x 8
##    skim_variable n_missing complete_rate   min   max empty n_unique whitespace
##  * <chr>             <int>         <dbl> <int> <int> <int>    <int>      <int>
##  1 many_ID               0         1         8    37     0      952          0
##  2 Title                 0         1         2   129     0      862          0
##  3 Beg                  81         0.918     4     4     0       70          0
##  4 D                     0         1         1     3     0        2          0
##  5 L                     0         1         1     3     0        4          0
##  6 J                     0         1         1     1     0        2          0
##  7 Signature            81         0.918     4     4     0       70          0
##  8 Force               248         0.749     4     4     0       66          0
##  9 treaty_ID             0         1         8    37     0      957          0
## 10 DESTA_ID              0         1         1     4     0      927          0
## 11 WTO                  55         0.944     1     1     0        2          0
## 
## $TREND
## -- Data Summary ------------------------
##                            Values
## Name                       X[[i]]
## Number of rows             730   
## Number of columns          7     
## _______________________          
## Column type frequency:           
##   character                7     
## ________________________         
## Group variables            None  
## 
## -- Variable type: character --------------------------------------------------------------------------------------------------------------------------------------------------------
## # A tibble: 7 x 8
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## * <chr>             <int>         <dbl> <int> <int> <int>    <int>      <int>
## 1 many_ID               0         1         8    37     0      723          0
## 2 Title                 0         1         6    93     0      728          0
## 3 Beg                   5         0.993     4     4     0       68          0
## 4 Signature             5         0.993     4     4     0       68          0
## 5 Force                 5         0.993     4     4     0       68          0
## 6 treaty_ID             0         1         8    37     0      729          0
## 7 TREND_ID              0         1         1     5     0      724          0
## 
## $LABPTA
## -- Data Summary ------------------------
##                            Values
## Name                       X[[i]]
## Number of rows             483   
## Number of columns          7     
## _______________________          
## Column type frequency:           
##   character                7     
## ________________________         
## Group variables            None  
## 
## -- Variable type: character --------------------------------------------------------------------------------------------------------------------------------------------------------
## # A tibble: 7 x 8
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## * <chr>             <int>         <dbl> <int> <int> <int>    <int>      <int>
## 1 many_ID               0             1    10    37     0      480          0
## 2 Title                 0             1     6    89     0      468          0
## 3 Beg                   0             1     4     4     0       27          0
## 4 Signature             0             1     4     4     0       27          0
## 5 Force                 0             1     4     4     0       27          0
## 6 treaty_ID             0             1    10    37     0      483          0
## 7 LABPTA_ID             0             1     1     3     0      483          0
## 
## $GPTAD
## -- Data Summary ------------------------
##                            Values
## Name                       X[[i]]
## Number of rows             346   
## Number of columns          9     
## _______________________          
## Column type frequency:           
##   character                9     
## ________________________         
## Group variables            None  
## 
## -- Variable type: character --------------------------------------------------------------------------------------------------------------------------------------------------------
## # A tibble: 9 x 8
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## * <chr>             <int>         <dbl> <int> <int> <int>    <int>      <int>
## 1 many_ID               0         1         8    35     0      340          0
## 2 Title                 0         1         3    74     0      343          0
## 3 Beg                  30         0.913    10    10     0      286          0
## 4 D                     0         1         0     1     5        2          0
## 5 L                     0         1         0     3     5        5          0
## 6 Signature            53         0.847    10    10     0      266          0
## 7 Force                64         0.815    10    10     0      199          0
## 8 treaty_ID             0         1         8    35     0      340          0
## 9 GPTAD_ID              0         1         1     3     0      346          0
}
}
\keyword{datasets}
