% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements database documentation}
\format{
The agreements database is a list that contains the
following 4 datasets: DESTA, TREND, LABPTA, GPTAD.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{DESTA: }{A dataset with 959 observations and the following
10 variables: manyID, Title, Beg, AgreementType, DocType, GeogArea,
Signature, Force, treatyID, destaID.}
\item{TREND: }{A dataset with 729 observations and the following
7 variables: manyID, Title, Beg, Signature, Force, treatyID, trendID.}
\item{LABPTA: }{A dataset with 483 observations and the following
7 variables: manyID, Title, Beg, Signature, Force, treatyID, labptaID.}
\item{GPTAD: }{A dataset with 340 observations and the following
10 variables: manyID, Title, Beg, AgreementType, DocType, GeogArea,
Signature, Force, treatyID, gptadID.}
}
}
\usage{
agreements
}
\description{
agreements database documentation
}
\details{
\preformatted{## $DESTA
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             959   
## Number of columns          10    
## _______________________          
## Column type frequency:           
##   character                10    
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ───────────────────────────────────────────────────────────────────────────────────────
##    skim_variable n_missing complete_rate   min   max empty n_unique whitespace
##  1 manyID                0         1         8    37     0      952          0
##  2 Title                 0         1         2   129     0      862          0
##  3 Beg                  81         0.916    10    10     0       70          0
##  4 AgreementType         0         1         1     3     0        2          0
##  5 DocType               0         1         1     1     0        3          0
##  6 GeogArea              0         1         1     1     0        2          0
##  7 Signature            81         0.916    10    10     0       70          0
##  8 Force               242         0.748    10    10     0       66          0
##  9 treatyID              0         1         8    37     0      957          0
## 10 destaID               0         1         1     4     0      926          0
## 
## $TREND
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             729   
## Number of columns          7     
## _______________________          
## Column type frequency:           
##   character                7     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ───────────────────────────────────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 manyID                0         1         8    37     0      723          0
## 2 Title                 0         1         6    93     0      728          0
## 3 Beg                   5         0.993    10    10     0       68          0
## 4 Signature             5         0.993    10    10     0       68          0
## 5 Force                 5         0.993    10    10     0       68          0
## 6 treatyID              0         1         8    37     0      729          0
## 7 trendID               0         1         1     5     0      723          0
## 
## $LABPTA
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             483   
## Number of columns          7     
## _______________________          
## Column type frequency:           
##   character                7     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ───────────────────────────────────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 manyID                0             1    10    37     0      480          0
## 2 Title                 0             1     6    89     0      468          0
## 3 Beg                   0             1    10    10     0       27          0
## 4 Signature             0             1    10    10     0       27          0
## 5 Force                 0             1    10    10     0       27          0
## 6 treatyID              0             1    10    37     0      483          0
## 7 labptaID              0             1     1     3     0      483          0
## 
## $GPTAD
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             340   
## Number of columns          10    
## _______________________          
## Column type frequency:           
##   character                10    
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ───────────────────────────────────────────────────────────────────────────────────────
##    skim_variable n_missing complete_rate   min   max empty n_unique whitespace
##  1 manyID                0         1         8    37     0      338          0
##  2 Title                 0         1         3    74     0      339          0
##  3 Beg                  30         0.912    10    10     0      286          0
##  4 AgreementType         0         1         0     1     5        2          0
##  5 DocType               0         1         0     1     5        4          0
##  6 GeogArea             75         0.779     0     1     5        3          0
##  7 Signature            53         0.844    10    10     0      266          0
##  8 Force                63         0.815    10    10     0      199          0
##  9 treatyID              0         1         8    35     0      340          0
## 10 gptadID               0         1         1     3     0      340          0
}
}
\keyword{datasets}
