% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_countryname.R
\name{code_countryname}
\alias{code_countryname}
\title{Code Country Names}
\usage{
code_countryname(name)
}
\arguments{
\item{name}{A character vector of country name}
}
\value{
A character vector of country name
}
\description{
Standardize country names from other languages.
}
\examples{
name <- c("argentine", "albanie", "Bélarus")
code_countryname(name)
}
