% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-gnevar.R
\docType{data}
\name{gnevar}
\alias{gnevar}
\title{gnevar database documentation}
\format{
The gnevar database is a list that contains the
following 1 datasets: GNEVAR.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{GNEVAR: }{A dataset with 4447 observations and the following
16 variables: manyID, Title, Beg, AgreementType, DocType, GeogArea,
Signature, Force, treatyID, destaID, gptadID, labptaID, trendID, totaID,
Memb.conditions, Memb.procedures.}
}
}
\usage{
gnevar
}
\description{
gnevar database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{## $GNEVAR
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             4447  
## Number of columns          16    
## _______________________          
## Column type frequency:           
##   character                13    
##   messydt                  3     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ────────────────────────────────────────────────────
##    skim_variable   n_missing complete_rate min max empty n_unique whitespace
##  1 manyID                  0         1       8  37     0     1393          0
##  2 Title                   0         1       2 129     0     2042          0
##  3 AgreementType        2609         0.413   0   3     7        3          0
##  4 DocType              2609         0.413   0   1     7        4          0
##  5 GeogArea             2742         0.383   0   1     7        4          0
##  6 treatyID                0         1       8  37     0     1515          0
##  7 destaID              3107         0.301   1   4     0      926          0
##  8 gptadID              3949         0.112   1   3     0      340          0
##  9 labptaID             3632         0.183   1   3     0      483          0
## 10 trendID              3351         0.246   1   5     0      723          0
## 11 totaID               3749         0.157   1   3     0      450          0
## 12 Memb.conditions         6         0.999   2  50     0       57          0
## 13 Memb.procedures         6         0.999   2  40     0       11          0
## 
## ── Variable type: messydt ──────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate n_empty n_unique max        min       
## 1 Beg                 122         0.973       0      610 2020-01-01 1947-01-01
## 2 Signature           147         0.967       0      597 2020-01-01 1947-01-01
## 3 Force               326         0.927       0      484 2020-01-01 1947-01-01
}\if{html}{\out{</div>}}
}
\keyword{datasets}
