% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships database documentation}
\format{
The memberships database is a list that contains the
following 2 datasets: GPTAD_MEM, DESTA_MEM.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{GPTAD_MEM: }{A dataset with 2221 observations and the following
9 variables: manyID, CountryID, Title, Beg, Signature, Force, Country, GPTAD_ID, treatyID.}
\item{DESTA_MEM: }{A dataset with 7772 observations and the following
9 variables: manyID, CountryID, Title, Beg, Signature, Force, Country, DESTA_ID, treatyID.}
}
}
\usage{
memberships
}
\description{
memberships database documentation
}
\details{
\subsection{NA}{

NA
}
}
\keyword{datasets}
