% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships database documentation}
\format{
The memberships database is a list that contains the
following 2 datasets: GPTAD_MEM, DESTA_MEM.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{GPTAD_MEM: }{A dataset with 2201 observations and the following
9 variables: manyID, CountryID, Title, Beg, Signature, Force,
CountryName, gptadID, treatyID.}
\item{DESTA_MEM: }{A dataset with 7492 observations and the following
9 variables: manyID, CountryID, Title, Beg, Signature, Force,
CountryName, destaID, treatyID.}
}
}
\usage{
memberships
}
\description{
memberships database documentation
}
\details{
\preformatted{## $GPTAD_MEM
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             2201  
## Number of columns          9     
## _______________________          
## Column type frequency:           
##   character                6     
##   messydt                  3     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ───────────────────────────────────────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 manyID                0         1         8    37     0      345          0
## 2 CountryID           101         0.954     3     3     0      214          0
## 3 Title                 0         1         3    74     0      345          0
## 4 CountryName           0         1         3    48     0      249          0
## 5 gptadID               0         1         1     3     0      346          0
## 6 treatyID              0         1         8    37     0      345          0
## 
## ── Variable type: messydt ─────────────────────────────────────────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate n_empty n_unique
## 1 Beg                 100         0.955       0      291
## 2 Signature           331         0.850       0      271
## 3 Force               285         0.871       0      200
## 
## $DESTA_MEM
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             7492  
## Number of columns          9     
## _______________________          
## Column type frequency:           
##   character                6     
##   messydt                  3     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ───────────────────────────────────────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 manyID                0         1        12    37     0     1164          0
## 2 CountryID            10         0.999     3     3     0      220          0
## 3 Title                 0         1         2   146     0     1090          0
## 4 CountryName           0         1         4    38     0      222          0
## 5 destaID               0         1         1     4     0      931          0
## 6 treatyID              0         1        12    37     0     1174          0
## 
## ── Variable type: messydt ─────────────────────────────────────────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate n_empty n_unique
## 1 Beg                 630         0.916       0       70
## 2 Signature           630         0.916       0       70
## 3 Force              2886         0.615       0       66
}
}
\keyword{datasets}
