% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships datacube documentation}
\format{
The memberships datacube is a list that contains the
following 3 datasets: GPTAD_MEM, DESTA_MEM, HUGGO_MEM.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{GPTAD_MEM: }{A dataset with 2192 observations and the following
9 variables: manyID, treatyID, Title, Begin, stateID, Signature, Force,
StateName, gptadID.}
\item{DESTA_MEM: }{A dataset with 7466 observations and the following
9 variables: manyID, treatyID, Title, Begin, stateID, Signature, Force,
StateName, destaID.}
\item{HUGGO_MEM: }{A dataset with 7202 observations and the following
15 variables: manyID, treatyID, Title, Begin, stateID, Signature, Force,
StateName, gptadID, destaID, StateRatification, StateSignature, StateForce,
StateEnd, DataCollection.}
}
}
\usage{
memberships
}
\description{
memberships datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $GPTAD_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2192|     0     |       0       |
#> |   treatyID    |  character  |            2192|     0     |       0       |
#> |     Title     |  character  |            2192|     0     |       0       |
#> |     Begin     |    mdate    |            2192|    99     |     4.52      |
#> |    stateID    |  character  |            2192|    55     |     2.51      |
#> |   Signature   |    mdate    |            2192|    310    |     14.14     |
#> |     Force     |    mdate    |            2192|    285    |      13       |
#> |   StateName   |  character  |            2192|    36     |     1.64      |
#> |    gptadID    |  character  |            2192|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $DESTA_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            7466|     0     |       0       |
#> |   treatyID    |  character  |            7466|     0     |       0       |
#> |     Title     |  character  |            7466|     0     |       0       |
#> |     Begin     |    mdate    |            7466|    630    |     8.44      |
#> |    stateID    |  character  |            7466|    11     |     0.15      |
#> |   Signature   |    mdate    |            7466|    630    |     8.44      |
#> |     Force     |    mdate    |            7466|   2864    |     38.36     |
#> |   StateName   |  character  |            7466|     0     |       0       |
#> |    destaID    |  character  |            7466|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO_MEM
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |            7202|     0     |       0       |
#> |    treatyID     |  character  |            7202|     0     |       0       |
#> |      Title      |  character  |            7202|     0     |       0       |
#> |      Begin      |    mdate    |            7202|    721    |     10.01     |
#> |     stateID     |  character  |            7202|     0     |       0       |
#> |    Signature    |    mdate    |            7202|    904    |     12.55     |
#> |      Force      |    mdate    |            7202|   1676    |     23.27     |
#> |    StateName    |  character  |            7202|     0     |       0       |
#> |     gptadID     |  character  |            7202|   5028    |     69.81     |
#> |     destaID     |  character  |            7202|   1384    |     19.22     |
#> |StateRatification|    mdate    |            7202|   6572    |     91.25     |
#> | StateSignature  |    mdate    |            7202|   5936    |     82.42     |
#> |   StateForce    |    mdate    |            7202|   6134    |     85.17     |
#> |    StateEnd     |    mdate    |            7202|   6913    |     95.99     |
#> | DataCollection  |  character  |            7202|    447    |     6.21      |
#> ------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
