% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships database documentation}
\format{
The memberships database is a list that contains the
following 3 datasets: GPTAD_MEM, DESTA_MEM, HUGGO_MEM.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{GPTAD_MEM: }{A dataset with 2192 observations and the following
9 variables: manyID, treatyID, Title, Beg, stateID, Signature, Force,
StateName, gptadID.}
\item{DESTA_MEM: }{A dataset with 7586 observations and the following
9 variables: manyID, treatyID, Title, Beg, stateID, Signature, Force,
StateName, destaID.}
\item{HUGGO_MEM: }{A dataset with 9078 observations and the following
11 variables: manyID, treatyID, Title, Beg, stateID, Signature, Force,
StateName, gptadID, destaID, changes.}
}
}
\usage{
memberships
}
\description{
memberships database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $GPTAD_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2192|     0     |       0       |
#> |   treatyID    |  character  |            2192|     0     |       0       |
#> |     Title     |  character  |            2192|     0     |       0       |
#> |      Beg      |    mdate    |            2192|    99     |     4.52      |
#> |    stateID    |  character  |            2192|    55     |     2.51      |
#> |   Signature   |    mdate    |            2192|    310    |     14.14     |
#> |     Force     |    mdate    |            2192|    285    |      13       |
#> |   StateName   |  character  |            2192|    36     |     1.64      |
#> |    gptadID    |  character  |            2192|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $DESTA_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            7586|     0     |       0       |
#> |   treatyID    |  character  |            7586|     0     |       0       |
#> |     Title     |  character  |            7586|     0     |       0       |
#> |      Beg      |    mdate    |            7586|    630    |      8.3      |
#> |    stateID    |  character  |            7586|    10     |     0.13      |
#> |   Signature   |    mdate    |            7586|    630    |      8.3      |
#> |     Force     |    mdate    |            7586|   2947    |     38.85     |
#> |   StateName   |  character  |            7586|     0     |       0       |
#> |    destaID    |  character  |            7586|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            9078|     0     |       0       |
#> |   treatyID    |  character  |            9078|     0     |       0       |
#> |     Title     |  character  |            9078|     0     |       0       |
#> |      Beg      |    mdate    |            9078|    726    |       8       |
#> |    stateID    |  character  |            9078|     0     |       0       |
#> |   Signature   |    mdate    |            9078|    916    |     10.09     |
#> |     Force     |    mdate    |            9078|   3146    |     34.66     |
#> |   StateName   |  character  |            9078|     0     |       0       |
#> |    gptadID    |  character  |            9078|   6925    |     76.28     |
#> |    destaID    |  character  |            9078|   1567    |     17.26     |
#> |    changes    |   numeric   |            9078|     0     |       0       |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
