% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships database documentation}
\format{
The memberships database is a list that contains the
following 2 datasets: GPTAD_MEM, DESTA_MEM.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{GPTAD_MEM: }{A dataset with 2221 observations and the following
9 variables: manyID, CountryID, Title, Beg, Signature, Force, CountryName, gptadID, treatyID.}
\item{DESTA_MEM: }{A dataset with 7772 observations and the following
9 variables: manyID, CountryID, Title, Beg, Signature, Force, CountryName, destaID, treatyID.}
}
}
\usage{
memberships
}
\description{
memberships database documentation
}
\details{
\preformatted{## $GPTAD_MEM
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             2221  
## Number of columns          9     
## _______________________          
## Column type frequency:           
##   character                9     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ──────────────────────────────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 manyID              100         0.955     8    37     0      315          0
## 2 CountryID           101         0.955     3     3     0      214          0
## 3 Title                 0         1         3    74     0      348          0
## 4 Beg                 100         0.955    10    10     0      291          0
## 5 Signature           332         0.851    10    10     0      271          0
## 6 Force               288         0.870    10    10     0      200          0
## 7 CountryName           0         1         3    48     0      249          0
## 8 gptadID               0         1         1     3     0      351          0
## 9 treatyID              0         1         8    37     0      345          0
## 
## $DESTA_MEM
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             7772  
## Number of columns          9     
## _______________________          
## Column type frequency:           
##   character                9     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ──────────────────────────────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 manyID              631         0.919    12    37     0     1082          0
## 2 CountryID            11         0.999     3     3     0      220          0
## 3 Title                 0         1         2   146     0     1090          0
## 4 Beg                 631         0.919     4     4     0       70          0
## 5 Signature           631         0.919     4     4     0       70          0
## 6 Force              2960         0.619     4     4     0       66          0
## 7 CountryName           0         1         4    38     0      222          0
## 8 destaID               0         1         1     4     0      931          0
## 9 treatyID              0         1        12    37     0     1174          0
}
}
\keyword{datasets}
