% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships database documentation}
\format{
The memberships database is a list that contains the
following 3 datasets: GPTAD_MEM, DESTA_MEM, HUGGO_MEM.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
}
\usage{
memberships
}
\description{
memberships database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $GPTAD_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            2198|     0     |       0       |
#> |    stateID    |  character  |            2198|    130    |     5.91      |
#> |     Title     |  character  |            2198|     0     |       0       |
#> |      Beg      |    mdate    |            2198|    99     |      4.5      |
#> |   Signature   |    mdate    |            2198|    328    |     14.92     |
#> |     Force     |    mdate    |            2198|    283    |     12.88     |
#> |   StateName   |  character  |            2198|     0     |       0       |
#> |    gptadID    |  character  |            2198|     0     |       0       |
#> |   treatyID    |  character  |            2198|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $DESTA_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            7492|     0     |       0       |
#> |    stateID    |  character  |            7492|    10     |     0.13      |
#> |     Title     |  character  |            7492|     0     |       0       |
#> |      Beg      |    mdate    |            7492|    630    |     8.41      |
#> |   Signature   |    mdate    |            7492|    630    |     8.41      |
#> |     Force     |    mdate    |            7492|   2886    |     38.52     |
#> |   StateName   |  character  |            7492|     0     |       0       |
#> |    destaID    |  character  |            7492|     0     |       0       |
#> |   treatyID    |  character  |            7492|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |            8973|     0     |       0       |
#> |    stateID    |  character  |            8973|     0     |       0       |
#> |     Title     |  character  |            8973|     0     |       0       |
#> |      Beg      |    mdate    |            8973|    720    |     8.02      |
#> |   Signature   |    mdate    |            8973|    932    |     10.39     |
#> |     Force     |    mdate    |            8973|   3121    |     34.78     |
#> |   StateName   |  character  |            8973|     0     |       0       |
#> |   treatyID    |  character  |            8973|     0     |       0       |
#> |    gptadID    |  character  |            8973|   6905    |     76.95     |
#> |    destaID    |  character  |            8973|   1491    |     16.62     |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
