% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qTrade-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships database documentation}
\format{
The memberships database is a list that contains the
following 2 datasets: GPTAD_MEM, DESTA_MEM.
For more information and references to each of the datasets used,
please use the \code{data_source()}, \code{data_contrast()}, and \code{data_evolution()}
functions.
\describe{
\item{GPTAD_MEM: }{A dataset with 2221 observations and the following
8 variables: qID_ref, Country_ID, Country, Title, Beg, Signature, Force, qID.}
\item{DESTA_MEM: }{A dataset with 7772 observations and the following
8 variables: qID_ref, DESTA_ID, Country, Title, Beg, Signature, Force, qID.}
}
}
\usage{
memberships
}
\description{
memberships database documentation
}
\keyword{datasets}
