% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-references.R
\docType{data}
\name{references}
\alias{references}
\title{references database documentation}
\format{
The references database is a list that contains the
following 1 datasets: DESTA_REF.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{DESTA_REF: }{A dataset with 584 observations and the following
3 variables: treatyID1, treatyID2, RefType.}
}
}
\usage{
references
}
\description{
references database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{## $DESTA_REF
## -- Data Summary ------------------------
##                            Values
## Name                       X[[i]]
## Number of rows             584   
## Number of columns          3     
## _______________________          
## Column type frequency:           
##   character                3     
## ________________________         
## Group variables            None  
## 
## -- Variable type: character ----------------------------------------------------
##   skim_variable n_missing complete_rate min max empty n_unique whitespace
## 1 treatyID1             0             1   8  37     0      378          0
## 2 treatyID2             0             1   8  37     0      378          0
## 3 RefType               0             1   5  10     0        4          0
}\if{html}{\out{</div>}}
}
\keyword{datasets}
