% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-references.R
\docType{data}
\name{references}
\alias{references}
\title{references database documentation}
\format{
The references database is a list that contains the
following 1 datasets: DESTA_REF.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{DESTA_REF: }{A dataset with 1000 observations and the following
3 variables: many_ID1, RefType, many_ID2.}
}
}
\usage{
references
}
\description{
references database documentation
}
\details{
\subsection{NA}{

NA
}
}
\keyword{datasets}
