% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-references.R
\docType{data}
\name{references}
\alias{references}
\title{references database documentation}
\format{
The references database is a list that contains the
following 1 datasets: DESTA_REF.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{DESTA_REF: }{A dataset with 1000 observations and the following
3 variables: manyID1, RefType, manyID2.}
}
}
\usage{
references
}
\description{
references database documentation
}
\details{
\preformatted{## $DESTA_REF
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             1000  
## Number of columns          3     
## _______________________          
## Column type frequency:           
##   character                3     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ──────────────────────────────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 manyID1               6         0.994     8    37     0      866          0
## 2 RefType               0         1         0    10   812        3          0
## 3 manyID2             822         0.178     8    37     0       59          0
}
}
\keyword{datasets}
