% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-references.R
\docType{data}
\name{references}
\alias{references}
\title{references datacube documentation}
\format{
The references datacube is a list that contains the
following 1 datasets: DESTA_REF.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{DESTA_REF: }{A dataset with 584 observations and the following
3 variables: treatyID1, treatyID2, RefType.}
}
}
\usage{
references
}
\description{
references datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $DESTA_REF
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |   treatyID1   |  character  |             584|     0     |       0       |
#> |   treatyID2   |  character  |             584|     0     |       0       |
#> |    RefType    |  character  |             584|     0     |       0       |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
