% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-texts.R
\docType{data}
\name{texts}
\alias{texts}
\title{texts database documentation}
\format{
The texts database is a list that contains the
following 1 datasets: GNEVAR_TXT.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{GNEVAR_TXT: }{A dataset with 1503 observations and the following
8 variables: manyID, Title, Beg, Signature, Force, treatyID, TreatyText, url.}
}
}
\usage{
texts
}
\description{
texts database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{## $GNEVAR_TXT
## -- Data Summary ------------------------
##                            Values
## Name                       X[[i]]
## Number of rows             1503  
## Number of columns          8     
## _______________________          
## Column type frequency:           
##   character                7     
##   list                     1     
## ________________________         
## Group variables            None  
## 
## -- Variable type: character ----------------------------------------------------
##   skim_variable n_missing complete_rate min max empty n_unique whitespace
## 1 manyID                0         1       8  37     0     1385          0
## 2 Title                 0         1       3 129     0     1331          0
## 3 Beg                 116         0.923  10  10     0      560          0
## 4 Signature           135         0.910  10  10     0      548          0
## 5 Force               254         0.831  10  10     0      419          0
## 6 treatyID              0         1       8  37     0     1389          0
## 7 url                 823         0.452  52  86     0      627          0
## 
## -- Variable type: list ---------------------------------------------------------
##   skim_variable n_missing complete_rate n_unique min_length max_length
## 1 TreatyText          473         0.685      853          0       1953
}\if{html}{\out{</div>}}
}
\keyword{datasets}
