% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-texts.R
\docType{data}
\name{texts}
\alias{texts}
\title{texts database documentation}
\format{
The texts database is a list that contains the
following 1 datasets: AGR_TXT.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{AGR_TXT: }{A dataset with 1503 observations and the following
13 variables: manyID, Title, Beg, Signature, Force, totaID, gptadID,
trendID, labptaID, treatyID, destaID, TreatyText, url.}
}
}
\usage{
texts
}
\description{
texts database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{## $AGR_TXT
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             1503  
## Number of columns          13    
## _______________________          
## Column type frequency:           
##   character                9     
##   list                     1     
##   mdate                    3     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate min max empty n_unique whitespace
## 1 manyID                0         1       8  37     0     1385          0
## 2 Title                 0         1       3 129     0     1331          0
## 3 totaID             1018         0.323   1   3     0      450          0
## 4 gptadID            1119         0.255   1   3     0      340          0
## 5 trendID             668         0.556   1   5     0      723          0
## 6 labptaID            916         0.391   1   3     0      483          0
## 7 treatyID              0         1       8  37     0     1389          0
## 8 destaID             437         0.709   1   4     0      926          0
## 9 url                 823         0.452  52  86     0      627          0
## 
## ── Variable type: list ─────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate n_unique min_length max_length
## 1 TreatyText            0             1      854          0       1953
## 
## ── Variable type: mdate ────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate n_empty n_unique max        min       
## 1 Beg                 116         0.923       0      560 2020-01-01 1947-01-01
## 2 Signature           135         0.910       0      548 2020-01-01 1947-01-01
## 3 Force               254         0.831       0      419 2020-01-01 1947-01-01
}\if{html}{\out{</div>}}
}
\keyword{datasets}
