% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-texts.R
\docType{data}
\name{texts}
\alias{texts}
\title{texts database documentation}
\format{
The texts database is a list that contains the
following 1 datasets: GNEVAR_TXT.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{GNEVAR_TXT: }{A dataset with 1440 observations and the following
13 variables: manyID, Title, Beg, Signature, Force, totaID, gptadID, destaID,
labptaID, trendID, treatyID, TreatyText, url.}
}
}
\usage{
texts
}
\description{
texts database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{## $GNEVAR_TXT
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |    manyID     |  character  |            1440|     0     |       0       |
## |     Title     |  character  |            1440|     0     |       0       |
## |      Beg      |    mdate    |            1440|    116    |     8.06      |
## |   Signature   |    mdate    |            1440|    135    |     9.38      |
## |     Force     |    mdate    |            1440|    253    |     17.57     |
## |    totaID     |  character  |            1440|    959    |     66.6      |
## |    gptadID    |  character  |            1440|   1058    |     73.47     |
## |    destaID    |  character  |            1440|    434    |     30.14     |
## |   labptaID    |  character  |            1440|    913    |     63.4      |
## |    trendID    |  character  |            1440|    665    |     46.18     |
## |   treatyID    |  character  |            1440|     0     |       0       |
## |  TreatyText   |    list     |            1440|    536    |     37.22     |
## |      url      |  character  |            1440|    766    |     53.19     |
## ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
