% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-texts.R
\docType{data}
\name{texts}
\alias{texts}
\title{texts database documentation}
\format{
The texts database is a list that contains the
following dataset: TOTA_TXT.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{TOTA_TXT: }{A dataset with 450 observations and the following
8 variables: manyID, Title, Beg, Signature, Force, TreatyText, treatyID, url.}
}
}
\usage{
texts
}
\description{
texts database documentation
}
\details{
\preformatted{## $TOTA_TXT
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             450   
## Number of columns          8     
## _______________________          
## Column type frequency:           
##   character                7     
##   list                     1     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ──────────────────────────────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 manyID                0             1    10    37     0      448          0
## 2 Title                 0             1     4    91     0      448          0
## 3 Beg                   0             1    10    10     0      407          0
## 4 Signature             0             1    10    10     0      407          0
## 5 Force                 0             1    10    12     0      318          0
## 6 treatyID              0             1    10    27     0      449          0
## 7 url                   0             1    64    66     0      450          0
## 
## ── Variable type: list ───────────────────────────────────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate n_unique min_length max_length
## 1 TreatyText            0             1      450          1          1
}
}
\keyword{datasets}
