% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-texts.R
\docType{data}
\name{texts}
\alias{texts}
\title{texts database documentation}
\format{
The texts database is a list that contains the
following 2 datasets: AGR_TXT, TOTA_TXT.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{AGR_TXT: }{A dataset with 1264 observations and the following
11 variables: manyID, gptadID, destaID, trendID, Title, Beg, Signature, Force, treatyID, labptaID, TreatyText.}
\item{TOTA_TXT: }{A dataset with 450 observations and the following
8 variables: manyID, Title, Beg, Signature, Force, TreatyText, treatyID, url.}
}
}
\usage{
texts
}
\description{
texts database documentation
}
\details{
\preformatted{## $AGR_TXT
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             1264  
## Number of columns          11    
## _______________________          
## Column type frequency:           
##   character                7     
##   list                     1     
##   messydt                  3     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 manyID                0         1         8    37     0     1208          0
## 2 gptadID             880         0.304     1     3     0      340          0
## 3 destaID             256         0.797     1     4     0      926          0
## 4 trendID             487         0.615     1     5     0      723          0
## 5 Title                 0         1         2   129     0     1081          0
## 6 treatyID              0         1         8    37     0     1212          0
## 7 labptaID            736         0.418     1     3     0      483          0
## 
## ── Variable type: list ─────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate n_unique min_length max_length
## 1 TreatyText            0             1      661          0       2605
## 
## ── Variable type: messydt ──────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate n_empty n_unique
## 1 Beg                 116         0.908       0      350
## 2 Signature           137         0.892       0      337
## 3 Force               255         0.798       0      246
## 
## $TOTA_TXT
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             450   
## Number of columns          8     
## _______________________          
## Column type frequency:           
##   character                4     
##   list                     1     
##   messydt                  3     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## 1 manyID                0             1    10    37     0      448          0
## 2 Title                 0             1     4    91     0      448          0
## 3 treatyID              0             1    10    27     0      449          0
## 4 url                   0             1    64    66     0      450          0
## 
## ── Variable type: list ─────────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate n_unique min_length max_length
## 1 TreatyText            0             1      450          1          1
## 
## ── Variable type: messydt ──────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate n_empty n_unique
## 1 Beg                   0             1       0      407
## 2 Signature             0             1       0      407
## 3 Force                 0             1       0      318
}
}
\keyword{datasets}
