% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manytrade-texts.R
\docType{data}
\name{texts}
\alias{texts}
\title{texts database documentation}
\format{
The texts database is a list that contains the
following 1 datasets: AGR_TXT.
To avoid overlaps in texts stored, AGR_TXT is based on a consolidated
version of the agreements database, to avoid duplicate entries.
Texts from the TOTA database (Alschner, Seiermann, and Skougarevskiy, 2017)
are given precedence.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{AGR_TXT: }{A dataset with 1740 observations and the following
13 variables: manyID, Title, Beg, Signature, Force, treatyID, totaID,
gptadID, trendID, labptaID, destaID, TreatyText, url.}
}
}
\usage{
texts
}
\description{
texts database documentation
}
\details{
\if{html}{\out{<div class="sourceCode {r, echo = FALSE, warning = FALSE}">}}\preformatted{lapply(texts, skimr::skim_without_charts)
}\if{html}{\out{</div>}}
}
\keyword{datasets}
